efc8d29 fix some window 64bit warnings

+++ smtk/Qt/qtAttribute.cxx
@@ -148,7 +148,7 @@ void qtAttribute::updateItemsData()
   std::size_t i, n = att->numberOfItems();
   for (i = 0; i < n; i++)
     {
-    qItem = this->createItem(att->item(i), this->Widget);
+    qItem = this->createItem(att->item(static_cast<int>(i)), this->Widget);
     if(qItem)
       {
       layout->addWidget(qItem->widget());
